Show logout template after logging out
