+++ lib/livebook_web/controllers/user_auth_html.ex
@@ -1,0 +1,5 @@
+defmodule LivebookWeb.UserAuthHTML do
+  use LivebookWeb, :html
+
+  embed_templates "user_auth_html/*"
+end
